+++ src/simulation/events.jl
@@ -5,7 +5,7 @@ using Random
 
 # General actions
 
-function mutantPathogen!(pathogen::Pathogen, class::Class)
+function mutantPathogen!(pathogen::Pathogen, population::Population)
     loci = rand(
         1:pathogen.type.num_loci, zeroTruncatedPoisson(pathogen.mean_mutations_per_replication)
     )
@@ -15,14 +15,14 @@ function mutantPathogen!(pathogen::Pathogen, class::Class)
     end
     seq = seq * pathogen.sequence[length(seq)+1:end]
 
-    if haskey(class.pathogens, seq)
-        return class.pathogens[seq]
+    if haskey(population.pathogens, seq)
+        return population.pathogens[seq]
     else
-        return newPathogen!(seq, class, pathogen.type)
+        return newPathogen!(seq, population, pathogen.type)
     end
 end
 
-function recombinantPathogens!(pathogen_1::Pathogen, pathogen_2::Pathogen, class::Class)
+function recombinantPathogens!(pathogen_1::Pathogen, pathogen_2::Pathogen, population::Population)
     children = MVector(pathogen_1.sequence, pathogen_2.sequence)
 
     if pathogen_1.mean_recombination_crossovers > 0 && pathogen_2.mean_recombination_crossovers > 0
@@ -56,88 +56,86 @@ function recombinantPathogens!(pathogen_1::Pathogen, pathogen_2::Pathogen, class
 
     # return SA[class.pathogens[children[1]], class.pathogens[children[2]]]
 
-    if !haskey(class.pathogens, children[1])
-        newPathogen!(children[1], class, pathogen_1.type)
+    if !haskey(population.pathogens, children[1])
+        newPathogen!(children[1], population, pathogen_1.type)
     end
-    return class.pathogens[children[1]]
+    return population.pathogens[children[1]]
 end
 
-function addPathogenToHost!(pathogen::Pathogen, host_idx::Int64, class::Class, population::Population, model::Model)
-    push!(class.hosts[host_idx].pathogens, pathogen)
-    push!(class.hosts[host_idx].pathogen_fractions, 0.0)
-    class.hosts[host_idx].pathogen_weights = catCol(
-        class.hosts[host_idx].pathogen_weights, zeros(Float64, NUM_PATHOGEN_EVENTS)
+function addPathogenToHost!(pathogen::Pathogen, host_idx::Int64, population::Population, model::Model)
+    push!(population.hosts[host_idx].pathogens, pathogen)
+    push!(population.hosts[host_idx].pathogen_fractions, 0.0)
+    population.hosts[host_idx].pathogen_weights = catCol(
+        population.hosts[host_idx].pathogen_weights, zeros(Float64, NUM_PATHOGEN_EVENTS)
     )
 
-    hostWeights!(host_idx, class, population, model)
+    hostWeights!(host_idx, population, model)
 end
 
-function addResponseToHost!(response::Response, host_idx::Int64, class::Class, population::Population, model::Model)
-    push!(class.hosts[host_idx].responses, response)
-    class.hosts[host_idx].response_weights = catCol(
-        class.hosts[host_idx].response_weights, zeros(Float64, NUM_RESPONSE_EVENTS)
+function addResponseToHost!(response::Response, host_idx::Int64, population::Population, model::Model)
+    push!(population.hosts[host_idx].responses, response)
+    population.hosts[host_idx].response_weights = catCol(
+        population.hosts[host_idx].response_weights, zeros(Float64, NUM_RESPONSE_EVENTS)
     )
 
-    hostWeights!(host_idx, class, population, model)
+    hostWeights!(host_idx, population, model)
 end
 
-function removePathogenFromHost!(pathogen_idx::Int64, host_idx::Int64, class::Class, population::Population, model::Model)
-    deleteat!(class.hosts[host_idx].pathogens, pathogen_idx)
-    deleteat!(class.hosts[host_idx].pathogen_fractions, pathogen_idx)
-    class.hosts[host_idx].pathogen_weights = class.hosts[host_idx].pathogen_weights[:, begin:end.!=pathogen_idx]
+function removePathogenFromHost!(pathogen_idx::Int64, host_idx::Int64, population::Population, model::Model)
+    deleteat!(population.hosts[host_idx].pathogens, pathogen_idx)
+    deleteat!(population.hosts[host_idx].pathogen_fractions, pathogen_idx)
+    population.hosts[host_idx].pathogen_weights = population.hosts[host_idx].pathogen_weights[:, begin:end.!=pathogen_idx]
 
-    hostWeights!(host_idx, class, population, model)
+    hostWeights!(host_idx, population, model)
 end
 
-function removeResponseFromHost!(response_idx::Int64, host_idx::Int64, class::Class, population::Population, model::Model)
-    deleteat!(class.hosts[host_idx].responses, response_idx)
-    class.hosts[host_idx].response_weights = class.hosts[host_idx].response_weights[:, begin:end.!=response_idx]
+function removeResponseFromHost!(response_idx::Int64, host_idx::Int64, population::Population, model::Model)
+    deleteat!(population.hosts[host_idx].responses, response_idx)
+    population.hosts[host_idx].response_weights = population.hosts[host_idx].response_weights[:, begin:end.!=response_idx]
 
-    hostWeights!(host_idx, class, population, model)
+    hostWeights!(host_idx, population, model)
 end
 
 function attemptInfection!(pathogen::Pathogen,
-        host_idx::Int64, class_idx::Int64, pop_idx::Int64, model::Model)
+        host_idx::Int64, pop_idx::Int64, model::Model)
     if !(
         pathogen in
-        model.populations[pop_idx].classes[class_idx].hosts[host_idx].pathogens
+        model.populations[pop_idx].hosts[host_idx].pathogens
         ) &&
         rand() < infectionProbability(
             pathogen,
-            model.populations[pop_idx].classes[class_idx].hosts[host_idx]
+            model.populations[pop_idx].hosts[host_idx]
         )
 
         addPathogenToHost!(
-            pathogen,
-            host_idx, model.populations[pop_idx].classes[class_idx],
-            model.populations[pop_idx], model
+            pathogen, host_idx, model.populations[pop_idx], model
         )
     end
 end
 
 
-function hostContact!(model::Model, rand_n::Float64, same_population::Bool)
-    host_idx_1, class_idx_1, pop_idx_1, rand_n = chooseHost(INTRA_POPULATION_CONTACT, model, rand_n)
+function hostContact!(model::Model, rand_n::Float64)
+    host_idx_1, pop_idx_1, rand_n = chooseHost(CONTACT, model, rand_n)
     pop_idx_2 = pop_idx_1
-    if !same_population
-        pop_idx_2, rand_n = choosePopulation(RECEIVE_CONTACT, model, rand_n)
-        #TODO: this sampling needs to be done according to contact rates from pop1
-    end
-    class_idx_2, rand_n = chooseClass(pop_idx_2, RECEIVE_CONTACT, model, rand_n)
-    host_idx_2, rand_n = chooseHost(class_idx_2, pop_idx_2, RECEIVE_CONTACT, model, rand_n)
+    # if !same_population
+    #     pop_idx_2, rand_n = choosePopulation(RECEIVE_CONTACT, model, rand_n)
+    #     #TODO: this sampling needs to be done according to contact rates from pop1
+    # end
+    # class_idx_2, rand_n = chooseClass(pop_idx_2, RECEIVE_CONTACT, model, rand_n)
+    host_idx_2, rand_n = chooseHost(pop_idx_2, RECEIVE_CONTACT, model, rand_n)
 
-    if host_idx_1 != host_idx_2 || class_idx_1 != class_idx_2 || pop_idx_1 != pop_idx_2
-        inocula = MVector{length(model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens)}([
+    if host_idx_1 != host_idx_2 || pop_idx_1 != pop_idx_2
+        inocula = MVector{length(model.populations[pop_idx_1].hosts[host_idx_1].pathogens)}([
             pois_rand(
-                model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx].mean_effective_inoculum *
-                model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogen_fractions[p_idx]
+                model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx].mean_effective_inoculum *
+                model.populations[pop_idx_1].hosts[host_idx_1].pathogen_fractions[p_idx]
             )
-            for p_idx in 1:length(model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens)
+            for p_idx in 1:length(model.populations[pop_idx_1].hosts[host_idx_1].pathogens)
         ])
         if !any(inocula .!= 0)
             idx, rand_n = randChoose(
                 rand_n,
-                model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogen_fractions,
+                model.populations[pop_idx_1].hosts[host_idx_1].pathogen_fractions,
                 1.0, regenerate_rand=true
             )
             inocula[idx] += 1
@@ -147,15 +145,15 @@ function hostContact!(model::Model, rand_n::Float64, same_population::Bool)
             if inocula[p_idx] > 0
                 mut_prob = min(
                     1.0, 1.0-exp(
-                        -model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx].mean_mutations_per_replication
+                        -model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx].mean_mutations_per_replication
                     )
                 )
                 # probability from Poisson PMF with k=0
                 rec_prob = 0.0
-                if length(model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens) > 1
+                if length(model.populations[pop_idx_1].hosts[host_idx_1].pathogens) > 1
                     rec_prob = min(
                         1.0, 1.0-exp(
-                            -model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx].mean_recombination_crossovers
+                            -model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx].mean_recombination_crossovers
                         )
                     )
                     # probability from Poisson PMF with k=0
@@ -165,55 +163,55 @@ function hostContact!(model::Model, rand_n::Float64, same_population::Bool)
                 num_mut_rec = binomial(inocula[p_idx], mut_prob * rec_prob)
                 if inocula[p_idx] - num_mut - num_rec - num_mut_rec > 0
                     attemptInfection!(
-                        model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx],
-                        host_idx_2, class_idx_2, pop_idx_2, model
+                        model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx],
+                        host_idx_2, pop_idx_2, model
                     )
                 end
                 for _ in 1:num_mut
                     attemptInfection!(
                         mutantPathogen!(
-                            model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx],
-                            model.populations[pop_idx_1].classes[class_idx_1]
-                        ), host_idx_2, class_idx_2, pop_idx_2, model
+                            model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx],
+                            model.populations[pop_idx_1]
+                        ), host_idx_2, pop_idx_2, model
                     )
                 end
                 for _ in 1:num_rec
-                    p_idx_2 = choosePathogen(host_idx_1, class_idx_1, pop_idx_1, RECOMBINANT_ESTABLISHMENT, model, rand_n)
+                    p_idx_2 = choosePathogen(host_idx_1, pop_idx_1, RECOMBINANT_ESTABLISHMENT, model, rand_n)
 
                     if p_idx != p_idx_2 && (
-                        model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx].type ==
-                        model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx_2].type)
+                        model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx].type ==
+                        model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx_2].type)
 
                         attemptInfection!(
                             recombinantPathogens!(
-                                model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx],
-                                model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx_2],
-                                model.populations[pop_idx_1].classes[class_idx_1]
-                            ), host_idx_2, class_idx_2, pop_idx_2, model
+                                model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx],
+                                model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx_2],
+                                model.populations[pop_idx_1]
+                            ), host_idx_2, pop_idx_2, model
                         )
                     end
                 end
                 for _ in 1:num_mut_rec
-                    p_idx_2 = choosePathogen(host_idx_1, class_idx_1, pop_idx_1, RECOMBINANT_ESTABLISHMENT, model, rand_n)
+                    p_idx_2 = choosePathogen(host_idx_1, pop_idx_1, RECOMBINANT_ESTABLISHMENT, model, rand_n)
 
                     if p_idx != p_idx_2
                         recombinant = recombinantPathogens!(
-                            model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx],
-                            model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx_2],
-                            model.populations[pop_idx_1].classes[class_idx_1]
+                            model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx],
+                            model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx_2],
+                            model.populations[pop_idx_1]
                         )
 
                         attemptInfection!(
-                            mutantPathogen!(recombinant, model.populations[pop_idx_1].classes[class_idx_1]),
-                            host_idx_2, class_idx_2, pop_idx_2, model
+                            mutantPathogen!(recombinant, model.populations[pop_idx_1]),
+                            host_idx_2, pop_idx_2, model
                         )
                     else
                         attemptInfection!(
                             mutantPathogen!(
-                                model.populations[pop_idx_1].classes[class_idx_1].hosts[host_idx_1].pathogens[p_idx],
-                                model.populations[pop_idx_1].classes[class_idx_1]
+                                model.populations[pop_idx_1].hosts[host_idx_1].pathogens[p_idx],
+                                model.populations[pop_idx_1]
                             ),
-                            host_idx_2, class_idx_2, pop_idx_2, model
+                            host_idx_2, pop_idx_2, model
                         )
                     end
                 end
@@ -225,40 +223,38 @@ end
 # Model events
 
 function establishMutant!(model::Model, rand_n::Float64)
-    pathogen_idx, host_idx, class_idx, pop_idx = choosePathogen(MUTANT_ESTABLISHMENT, model, rand_n)
+    pathogen_idx, host_idx, pop_idx = choosePathogen(MUTANT_ESTABLISHMENT, model, rand_n)
 
     mut = mutantPathogen!(
-        model.populations[pop_idx].classes[class_idx].hosts[host_idx].pathogens[pathogen_idx],
-        model.populations[pop_idx].classes[class_idx]
+        model.populations[pop_idx].hosts[host_idx].pathogens[pathogen_idx],
+        model.populations[pop_idx]
     )
 
-    attemptInfection!(mut, host_idx, class_idx, pop_idx, model)
+    attemptInfection!(mut, host_idx, pop_idx, model)
 end
 
 function clearPathogen!(model::Model, rand_n::Float64)
-    pathogen_idx, host_idx, class_idx, pop_idx = choosePathogen(CLEARANCE, model, rand_n)
+    pathogen_idx, host_idx, pop_idx = choosePathogen(CLEARANCE, model, rand_n)
 
     removePathogenFromHost!(
         pathogen_idx, host_idx,
-        model.populations[pop_idx].classes[class_idx],
         model.populations[pop_idx], model
     )
 end
 
 function acquireResponse!(model::Model, rand_n::Float64)
-    pathogen_idx, host_idx, class_idx, pop_idx = choosePathogen(CLEARANCE, model, rand_n)
+    pathogen_idx, host_idx, pop_idx = choosePathogen(CLEARANCE, model, rand_n)
 
-    responses = model.populations[pop_idx].classes[class_idx].parameters.developResponses(
-        model.populations[pop_idx].classes[class_idx].hosts[host_idx].pathogens[pathogen_idx],
-        model.populations[pop_idx].classes[class_idx].hosts[host_idx],
-        model.populations[pop_idx].classes[class_idx],
+    responses = model.populations[pop_idx].parameters.developResponses(
+        model.populations[pop_idx].hosts[host_idx].pathogens[pathogen_idx],
+        model.populations[pop_idx].hosts[host_idx],
+        model.populations[pop_idx],
     )
 
     for response in responses
-        if !(response in model.populations[pop_idx].classes[class_idx].hosts[host_idx].responses)
+        if !(response in model.populations[pop_idx].hosts[host_idx].responses)
             addResponseToHost!(
                 response, host_idx,
-                model.populations[pop_idx].classes[class_idx],
                 model.populations[pop_idx], model
             )
         end
@@ -266,30 +262,21 @@ function acquireResponse!(model::Model, rand_n::Float64)
 end
 
 function establishRecombinant!(model::Model, rand_n::Float64)
-    pathogen_idx_1, host_idx, class_idx, pop_idx, rand_n = choosePathogen(RECOMBINANT_ESTABLISHMENT, model, rand_n)
-    pathogen_idx_2 = choosePathogen(host_idx, class_idx, pop_idx, RECOMBINANT_ESTABLISHMENT, model, rand_n)
+    pathogen_idx_1, host_idx, pop_idx, rand_n = choosePathogen(RECOMBINANT_ESTABLISHMENT, model, rand_n)
+    pathogen_idx_2 = choosePathogen(host_idx, pop_idx, RECOMBINANT_ESTABLISHMENT, model, rand_n)
 
     if pathogen_idx_1 != pathogen_idx_2 && pathogen_idx_1.type == pathogen_idx_2.type
         recombinant = recombinantPathogens!(
-            model.populations[pop_idx].classes[class_idx].hosts[host_idx].pathogens[pathogen_idx_1],
-            model.populations[pop_idx].classes[class_idx].hosts[host_idx].pathogens[pathogen_idx_2],
-            model.populations[pop_idx].classes[class_idx]
+            model.populations[pop_idx].hosts[host_idx].pathogens[pathogen_idx_1],
+            model.populations[pop_idx].hosts[host_idx].pathogens[pathogen_idx_2],
+            model.populations[pop_idx]
         )
 
-        attemptInfection!(recombinant, host_idx, class_idx, pop_idx, model)
+        attemptInfection!(recombinant, host_idx, pop_idx, model)
     end
-end
-
-function intraPopulationContact!(model::Model, rand_n::Float64)
-    hostContact!(model, rand_n, true)
-end
-
-function interPopulationContact!(model::Model, rand_n::Float64)
-    hostContact!(model, rand_n, false)
 end
 
 event_functions = SA[
     establishMutant!, clearPathogen!, acquireResponse!,
-    establishRecombinant!, intraPopulationContact!,
-    interPopulationContact!,
+    establishRecombinant!, hostContact!,
 ]