+++ src/simulation/initializers.jl
@@ -1,7 +1,7 @@
 using StaticArrays
 
-function newPathogen!(sequence::String, class::Class, type::PathogenType)
-    class.pathogens[sequence] = Pathogen(
+function newPathogen!(sequence::String, population::Population, type::PathogenType)
+    population.pathogens[sequence] = Pathogen(
         sequence, pathogenSequenceCoefficients(sequence, type),
         type.mean_effective_inoculum * type.inoculumCoefficient(sequence),
         type.mean_mutations_per_replication * type.mutationCoefficient(sequence),
@@ -9,36 +9,36 @@ function newPathogen!(sequence::String, class::Class, type::PathogenType)
         type
     )
 
-    return class.pathogens[sequence]
+    return population.pathogens[sequence]
 end
 
 function newResponse!(
     imprinted_pathogen::Pathogen, matured_pathogen::Pathogen, parent::Tuple{String,String,String},
-    class::Class, type::ResponseType)
-    class.responses[(imprinted_pathogen.sequence, matured_pathogen.sequence, type.id)] = Response(
+    population::Population, type::ResponseType)
+    population.responses[(imprinted_pathogen.sequence, matured_pathogen.sequence, type.id)] = Response(
         parent, imprinted_pathogen, matured_pathogen,
         responseStaticCoefficients(imprinted_pathogen.sequence, matured_pathogen.sequence, type),
         type
     )
 
-    return class.responses[(imprinted_pathogen.sequence, matured_pathogen.sequence, type.id)]
+    return population.responses[(imprinted_pathogen.sequence, matured_pathogen.sequence, type.id)]
 end
 
-function newHost!(class::Class, population::Population, model::Model)
-    push!(class.hosts, Host(
-        length(class.hosts) + 1,
+function newHost!(population::Population, model::Model)
+    push!(population.hosts, Host(
+        length(population.hosts) + 1,
         Vector{Pathogen}(undef, 0), Vector{Response}(undef, 0),
         Vector{Float64}(undef, 0),
         Matrix{Float64}(undef, NUM_PATHOGEN_EVENTS, 0),
         Matrix{Float64}(undef, NUM_RESPONSE_EVENTS, 0),
     ))
-    class.host_weights = catCol(class.host_weights, zeros(Float64, NUM_EVENTS))
-    class.host_weights_receive = catCol(class.host_weights_receive, zeros(Float64, NUM_CHOICE_MODIFIERS - 1))
+    population.host_weights = catCol(population.host_weights, zeros(Float64, NUM_EVENTS))
+    population.host_weights_receive = catCol(population.host_weights_receive, zeros(Float64, NUM_CHOICE_MODIFIERS - 1))
 
     propagateWeightChanges!(
-        - model.population_weights[INTRA_POPULATION_CONTACT, model.population_dict[population.id]] /
+        - model.population_weights[CONTACT, model.population_dict[population.id]] /
         (population.total_hosts + 1),
-        population, INTRA_POPULATION_CONTACT, model
+        population, CONTACT, model
     )
     #TODO: update inter-pop contacts
     population.total_hosts += 1
@@ -46,14 +46,14 @@ function newHost!(class::Class, population::Population, model::Model)
     for coef in EVENTS
         if START_COEFFICIENTS[coef] != 0
             propagateWeightChanges!(
-                START_COEFFICIENTS[coef], length(class.hosts), class, population, coef, model
+                START_COEFFICIENTS[coef], length(population.hosts), population, coef, model
             )
         end
     end
     for coef in CHOICE_MODIFIERS[begin:end-1]
         if START_COEFFICIENTS[coef] != 0
             propagateWeightReceiveChanges!(
-                START_COEFFICIENTS[coef], length(class.hosts), class, population, coef, model
+                START_COEFFICIENTS[coef], length(population.hosts), population, coef, model
             )
         end
     end
@@ -65,29 +65,29 @@ function newHost!(class::Class, population::Population, model::Model)
     #     , length(class.hosts), class, population, model
     # )
 
-    return class.hosts[end]
+    return population.hosts[end]
 end
 
-function newClass!(id::String, parameters::ClassParameters, population::Population)
-    push!(population.classes, Class(
+# function newClass!(id::String, parameters::ClassParameters, population::Population)
+#     push!(population.classes, Class(
+#         id, parameters,
+#         Dict{String,Pathogen}(), Dict{Tuple{String,String,String},Response}(),
+#         Vector{Host}(undef, 0),
+#         Matrix{Float64}(undef, NUM_EVENTS, 0),
+#         Matrix{Float64}(undef, NUM_CHOICE_MODIFIERS - 1, 0),
+#     ))
+#     population.class_dict[id] = length(population.classes)
+#     population.class_weights = catCol(population.class_weights, zeros(Float64, NUM_EVENTS))
+#     population.class_weights_receive = catCol(population.class_weights_receive, zeros(Float64, NUM_CHOICE_MODIFIERS - 1))
+
+#     return population.classes[end]
+# end
+
+function newPopulation!(id::String, parameters::PopulationParameters, model::Model)
+    push!(model.populations, Population(
         id, parameters,
         Dict{String,Pathogen}(), Dict{Tuple{String,String,String},Response}(),
         Vector{Host}(undef, 0),
-        Matrix{Float64}(undef, NUM_EVENTS, 0),
-        Matrix{Float64}(undef, NUM_CHOICE_MODIFIERS - 1, 0),
-    ))
-    population.class_dict[id] = length(population.classes)
-    population.class_weights = catCol(population.class_weights, zeros(Float64, NUM_EVENTS))
-    population.class_weights_receive = catCol(population.class_weights_receive, zeros(Float64, NUM_CHOICE_MODIFIERS - 1))
-
-    return population.classes[end]
-end
-
-function newPopulation!(id::String, model::Model)
-    push!(model.populations, Population(
-        id,
-        Vector{Class}(undef, 0),
-        Dict{String,Int64}(),
         Matrix{Float64}(undef, NUM_EVENTS, 0),
         Matrix{Float64}(undef, NUM_CHOICE_MODIFIERS - 1, 0),
         0, 0.0, 0.0